% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.R
\name{gsa_citypair}
\alias{gsa_citypair}
\title{Search City Pair Program airfares}
\usage{gsa_citypair(origin = NULL, destination = NULL, award_year = as.integer(format(Sys.Date(), "\%Y")), api_key = "DEMO_KEY")}
\arguments{
\item{origin}{Origin airport code}
\item{destination}{Destination airport code}
\item{award_year}{Fiscal year for awards}
\item{api_key}{API key}
}
\value{tibble with airfare details}
\description{Returns discounted City Pair Program airfares for federal travelers.}
