% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.R
\name{gsa_perdiem_state}
\alias{gsa_perdiem_state}
\title{List all per diem locations for a state}
\usage{gsa_perdiem_state(state, year = as.integer(format(Sys.Date(), "\%Y")), api_key = "DEMO_KEY")}
\arguments{
\item{state}{Two-letter state abbreviation}
\item{year}{Fiscal year}
\item{api_key}{API key}
}
\value{tibble with city, county, meals rate, and monthly lodging rates}
\description{Returns per diem rate areas defined for a state.}
