% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.R
\name{gsa_perdiem_zip}
\alias{gsa_perdiem_zip}
\title{Get per diem rates by zip code}
\usage{gsa_perdiem_zip(zip, year = as.integer(format(Sys.Date(), "\%Y")), api_key = "DEMO_KEY")}
\arguments{
\item{zip}{Five-digit zip code}
\item{year}{Fiscal year}
\item{api_key}{API key}
}
\value{tibble with per diem rate details}
\description{Get per diem rates by zip code.}
