% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.R
\name{gsa_site}
\alias{gsa_site}
\title{Get scan results for a specific website}
\usage{gsa_site(target_url, api_key = "DEMO_KEY")}
\arguments{
\item{target_url}{Full domain to look up}
\item{api_key}{API key}
}
\value{tibble with one row of scan fields}
\description{Get scan results for a specific website.}
